handle_portmap
